Safari-611.1.10/LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/interaction/focus/focus-input-type-switch-expected.txt

Can still focus on inputs that change types



FAIL Can change an input's type during focus handler without breaking focus assert_equals: The element was correctly focused expected Element node <input type="text" value="123" onfocus="javascript:event.... but got Element node <body><h1>Can still focus on inputs that change types</h1...
